requirements.txt
pyaudio
paho-mqtt
numpy